% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absorbance.R
\docType{data}
\name{absorbance}
\alias{absorbance}
\title{CDOM absorbance data.}
\format{A data frame with 711 rows and 4 variables}
\usage{
data("absorbance")
}
\description{
Simple absorbance spectra used to test package's functions.
}
\details{
\itemize{
  \item wavelength.  Wavelengths used for measurements (190-900 nm.)
  \item absorbance
}
}
\examples{

data("absorbance")

plot(absorbance$wavelength, absorbance$sample1,
  type = "l",
  xlab = "Wavelengths", ylab = "Absorbance per meter"
)
lines(absorbance$wavelength, absorbance$sample2, col = "blue")
lines(absorbance$wavelength, absorbance$sample3, col = "red")
}
\keyword{datasets}
