% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_bind}
\alias{eem_bind}
\title{Bind eem or eemlist}
\usage{
eem_bind(...)
}
\arguments{
\item{...}{One or more object of class \code{eemlist}.}
}
\value{
An object of \code{eemlist}.
}
\description{
Function to bind EEMs that have been loaded from different folders or have
been processed differently.
}
\examples{
file <- system.file("extdata/cary/eem/", "sample1.csv", package = "eemR")
eem <- eem_read(file)

eem <- eem_bind(eem, eem)
}

