% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_metrics.R
\name{eem_biological_index}
\alias{eem_biological_index}
\title{Calculate the biological fluorescence index (BIX)}
\usage{
eem_biological_index(eem)
}
\arguments{
\item{eem}{An object of class \code{eem} or \code{eemlist}.}
}
\value{
A data frame containing the biological index (BIX) for each eem.
}
\description{
The biological fluorescence index (BIX) is calculated by
  dividing the fluorescence at excitation 310 nm and emission at 380 nm (ex =
  310, em = 430) by that at excitation 310 nm and emission at 430 nm (ex =
  310, em = 380).
}
\examples{
file <- system.file("extdata/cary/eem/", package = "eemR")
eem <- eem_read(file)

eem_biological_index(eem)

}
\references{
Huguet, A., Vacher, L., Relexans, S., Saubusse, S., Froidefond,
  J. M., & Parlanti, E. (2009). Properties of fluorescent dissolved organic
  matter in the Gironde Estuary. Organic Geochemistry, 40(6), 706-719.

  \url{http://doi.org/10.1016/j.orggeochem.2009.03.002}
}

