% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_metrics.R
\name{eem_coble_peaks}
\alias{eem_coble_peaks}
\title{Extrace fluorescence peaks}
\usage{
eem_coble_peaks(eem)
}
\arguments{
\item{eem}{An object of class \code{eem}}
}
\value{
A data frame containing peaks B, T, A, M and C
}
\description{
Extrace fluorescence peaks
}
\examples{
file <- system.file("extdata/eem", "sample1.csv", package = "eem")
eem <- eem_read(file)

eem_coble_peaks(eem)
}

