% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_extract_blank}
\alias{eem_extract_blank}
\title{Extract blank EEM}
\usage{
eem_extract_blank(eem, average = TRUE)
}
\arguments{
\item{eem}{An object of class \code{eemlist}.}

\item{average}{Logical. If TRUE blank EEMs will be averaged}
}
\value{
An object of class \code{eemlist}.
}
\description{
Extract blank EEM
}
