% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_metrics.R
\name{eem_humification_index}
\alias{eem_humification_index}
\title{Calculate the humification index (HIX)}
\usage{
eem_humification_index(eem)
}
\arguments{
\item{eem}{An object of class \code{eem} or \code{eemlist}.}
}
\value{
A data frame containing the humification index (HIX) for each eem.
}
\description{
The fluorescence humification index (HIX), which compares two
  broad aromatic dominated fluorescence maxima, is calculated at 255 nm
  excitation by dividing the integrated emission from 435 to 480 nm by the
  integrated emission from 300 to 346 nm.
}
\examples{
file <- system.file("extdata/eem", package = "eemR")
eem <- eem_read(file)

eem_humification_index(eem)

}
\references{
Kalbitz, K., Geyer, W., & Geyer, S. (1999). Spectroscopic
  properties of dissolved humic substances - a reflection of land use history
  in a fen area. Biogeochemistry, 47(2), 219-238.

  \url{http://doi.org/10.1023/A:1006134214244}
}

