% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_names<-}
\alias{eem_names<-}
\title{Set the sample names of an eem or eemlist objects}
\usage{
eem_names(x) <- value
}
\arguments{
\item{x}{An object of class \code{eem} or \code{eemlist}.}

\item{value}{A character vector with new sample names. Must be equal
in length to the number of samples in the \code{eem} or \code{eemlist}.}
}
\value{
An \code{eem} or \code{eemlist}.
}
\description{
Set the sample names of an eem or eemlist objects
}
\examples{
folder <- system.file("extdata/cary/scans_day_1", package = "eemR")
eems <- eem_read(folder, import_function = "cary")

eem_names(eems)
eem_names(eems) <- c("a", "b", "c", "d")
eem_names(eems)
}
