% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_names}
\alias{eem_names}
\title{The names of an eem or eemlist objects}
\usage{
eem_names(eem)
}
\arguments{
\item{eem}{An object of class \code{eemlist}.}
}
\value{
An object of class \code{eemlist}.

A character vector containing the names of the EEMs.
}
\description{
The names of an eem or eemlist objects
}
\examples{
file <- system.file("extdata/cary/", package = "eemR")
eem <- eem_read(file, recursive = TRUE)

eem_names(eem)

}

