% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_parafac.R
\name{eem_plot_parafac}
\alias{eem_plot_parafac}
\title{Plot a PARAFAC model}
\usage{
eem_plot_parafac(x, type, ...)
}
\arguments{
\item{x}{A data frame returned by \code{eem_read_parafac}.}

\item{type}{Character. \emph{loading} for the loading plot or
\emph{component} for the PARAFAC model.}

\item{...}{Additional parameters (not used at the moment).}
}
\value{
A ggplot2 object
}
\description{
Plot a PARAFAC model
}
\examples{
f <- system.file("extdata/parafac_model.mat", package = "eemR")

m1 <- eem_read_parafac(f, object = "mymodel", 4)

p <- eem_plot_parafac(m1, type = "loading")
p

p <- eem_plot_parafac(m1, type = "component")
p
}

