% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_read.R
\name{eem_read}
\alias{eem_read}
\title{Read excitation-emission fluorecence matrix (eem)}
\usage{
eem_read(file)
}
\arguments{
\item{file}{File name or folder containing fluorescence file(s).}
}
\value{
If \code{file} is a single filename:

  An object of class \code{eem} containing: \itemize{ \item sample The file
  name of the eem. \item x A matrix with fluorescence values. \item em
  Emission vector of wavelengths. \item ex Excitation vector of wavelengths.
  }

  If \code{file} is a folder, the function returns an object of class
  \code{eemlist} which is simply a list of \code{eem}.
}
\description{
Read excitation-emission fluorecence matrix (eem)
}
\details{
At the moment,
  \href{https://www.agilent.com/en-us/products/fluorescence/fluorescence-systems/cary-eclipse-fluorescence-spectrophotometer}{Cary
   Eclipse} and
  \href{http://www.horiba.com/us/en/scientific/products/fluorescence-spectroscopy/}{Aqualog}
   EEMs are supported.

  \code{eemR} will automatically try to determine from which
  spectrofluorometer the files originate and load the data accordingly. Note
  that EEMs are reshaped so X[1, 1] represents the fluoresence intensity at
  X[min(ex), min(em)].
}
\examples{
file <- system.file("extdata/cary/eem/", "sample1.csv", package = "eemR")
eem <- eem_read(file)
}

