% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_parafac.R
\name{eem_read_parafac}
\alias{eem_read_parafac}
\title{Read a PARAFAC model}
\usage{
eem_read_parafac(matfile, object, ncomp)
}
\arguments{
\item{matfile}{Path of the mat file.}

\item{object}{Matlab object containing the model (ex. \emph{Test2}).}

\item{ncomp}{Numeric. The number of components to extract from \code{object}.}
}
\value{
A data frame with loading.
}
\description{
Read a PARAFAC model
}
\examples{
f <- system.file("extdata/parafac_model.mat", package = "eemR")
m1 <- eem_read_parafac(f, object = "mymodel", 4)
}

