% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove}
\alias{eem_remove}
\title{Remove EEM samples}
\usage{
eem_remove(eem, sample)
}
\arguments{
\item{eem}{Either an object of class \code{eem} or \code{eemlist}.}

\item{sample}{Either numeric of character vector. See \code{details} for more
information.}
}
\value{
Either an object of class \code{eem} or \code{eemlist}.
}
\description{
Remove EEM samples
}
\details{
\code{sample} argument can be either numeric or character vector. If
  it is numeric, samples at specified index will be removed.

  If \code{sample} is character, regular expression will be used and all
  sample names that have a partial or complete match with the expression will
  be removed. See \code{examples} for more details.
}
\examples{
folder <- system.file("extdata/cary/eem", package = "eemR")
eems <- eem_read(folder)

eem_remove(eems, c(1, 3)) ## Removes samples 1 and 3
eem_remove(eems, 2:3) ## Removes samples 2 and 3

## Remove ALL samples containing "3" in their names.
eem_remove(eems, "3")

## Remove ALL samples containing either character "s" or character "2" in their names.
eem_remove(eems, c("s", "2"))

## Remove ALL samples containing "blank" or "nano"
eem_remove(eems, c("blank", "nano"))

}

