% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove_scattering}
\alias{eem_remove_scattering}
\title{Remove Raman and Rayleigh scattering}
\usage{
eem_remove_scattering(eem, type, order = 1, width)
}
\arguments{
\item{eem}{Either an object of class \code{eem} or a list of \code{eem}.}

\item{type}{A string, either "raman" or "rayleigh"}

\item{order}{A integer number, either 1 (first order) or 2 (second order)}

\item{width}{Slit width in nm for the cut}
}
\value{
Either an object of class \code{eem} or a list of \code{eem}.
}
\description{
Remove Raman and Rayleigh scattering
}
\examples{
fluo <- system.file("extdata/eem", "sample1.csv", package = "eem")
b <- system.file("extdata", "nano.csv", package = "eem")
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)

eem <- eem_read(fluo, ex, em)

eem_remove_scattering(eem = eem, type = "raman", order = 1, width = 10)
}

