% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemr-package.R
\docType{package}
\name{eemr-package}
\alias{eemr}
\alias{eemr-package}
\title{eemr: Coherence-Based Expression Module Search}
\description{
R bindings (via Rcpp) for the eemCpp coherence-based expression module search tool. Given an expression matrix and a collection of candidate gene sets, eem_search() finds coherently expressed subsets of each gene set ("expression modules") and assigns them a significance p-value based on a permutation test.
}
\author{
\strong{Maintainer}: Atsushi Niida \email{atusiniida@gmail.com}

Authors:
\itemize{
  \item Atsushi Niida \email{atusiniida@gmail.com}
}

}
