% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{names<-.eem}
\alias{names<-.eem}
\title{Set the sample names of an eem object}
\usage{
\method{names}{eem}(x) <- value
}
\arguments{
\item{x}{An object of class \code{eem}.}

\item{value}{A character vector with new sample name.}
}
\value{
An \code{eem}.
}
\description{
Set the sample names of an eem object
}
\examples{
file <- system.file("extdata/cary/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)

names(eem)
names(eem) <- "a"
names(eem)

}

