% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{names<-.eemlist}
\alias{names<-.eemlist}
\title{Set the sample names of an eemlist object}
\usage{
\method{names}{eemlist}(x) <- value
}
\arguments{
\item{x}{An object of class \code{eemlist}.}

\item{value}{A character vector with new sample names. Must be equal in
length to the number of samples in the \code{eemlist}.}
}
\value{
An \code{eemlist}.
}
\description{
Set the sample names of an eemlist object
}
\examples{
folder <- system.file("extdata/cary/eem", package = "eemR")
eems <- eem_read(folder)

names(eems)
names(eems) <- c("a", "b", "c")
names(eems)

}

