% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{names.eem}
\alias{names.eem}
\title{The Names of an eem Object}
\usage{
\method{names}{eem}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eem}.}

\item{...}{Extra arguments.}
}
\value{
A character vector containing the name of the EEM.
}
\description{
The Names of an eem Object
}
\examples{
file <- system.file("extdata/cary/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)

names(eem)

}

