% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{names.eemlist}
\alias{names.eemlist}
\title{The Names of an eemlist Object}
\usage{
\method{names}{eemlist}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eem}.}

\item{...}{Extra arguments.}
}
\value{
A character vector containing the names of the EEMs.
}
\description{
The Names of an eemlist Object
}
\examples{
folder <- system.file("extdata/cary/eem", package = "eemR")
eems <- eem_read(folder)

names(eems)

}

