% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_parafac.R
\name{plot.parafac_model}
\alias{plot.parafac_model}
\title{Plot a PARAFAC model}
\usage{
\method{plot}{parafac_model}(x, nrow, ncol, ...)
}
\arguments{
\item{x}{A data frame returned by \code{eem_read_parafac}.}

\item{nrow}{Number of rows in the plot grid.}

\item{ncol}{Number of columns in the plot grid.}

\item{...}{Additional parameters (not used at the moment).}
}
\value{
A ggplot2 object
}
\description{
Plot a PARAFAC model
}
\details{
Use \code{vignette("parafac")} for more details.
}
\examples{

library(ggplot2)

f <- system.file("extdata/parafac_model.mat", package = "eemR")
m1 <- eem_read_parafac(f, object = "mymodel", 4)

p <- plot(m1, nrow = 2, ncol = 2)
p
}

