% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{summary.eem}
\alias{summary.eem}
\title{Display summary of an eem object}
\usage{
\method{summary}{eem}(object, ...)
}
\arguments{
\item{object}{An object of class \code{eem}}

\item{...}{Extra arguments}
}
\description{
Display summary of an eem object
}
\examples{
fluo <- system.file("extdata/eem", "sample1.csv", package = "eem")
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)
eem <- eem_read(fluo, ex, em)
summary(eem)
}
\references{
\url{http://linkinghub.elsevier.com/retrieve/pii/0304420395000623}
}

