% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{summary.eemlist}
\alias{summary.eemlist}
\title{Display summary of an eemlist object}
\usage{
\method{summary}{eemlist}(object, ...)
}
\arguments{
\item{object}{An object of class \code{eemlist}.}

\item{...}{Extra arguments.}
}
\description{
Display summary of an eemlist object
}
\examples{
folder <- system.file("extdata/cary/eem/", package = "eemR")
eem <- eem_read(folder)

summary(eem)
}

