% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_import.R
\name{eem}
\alias{eem}
\title{eem constructor}
\usage{
eem(data)
}
\arguments{
\item{data}{A list containing "file", "x", "em", "ex".}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
eem constructor
}
