% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_metrics.R
\name{eem_coble_peaks}
\alias{eem_coble_peaks}
\title{Extrace fluorescence peaks}
\usage{
eem_coble_peaks(eem)
}
\arguments{
\item{eem}{An object of class \code{eem}}
}
\value{
A data frame containing peaks B, T, A, M and C
}
\description{
Extrace fluorescence peaks
}
\examples{
fluo <- "/home/persican/Desktop/test/eem/sn21.csv"
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)
eem <- eem_read(fluo, ex, em)
eem_coble_peaks(eem)
}

