% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_metrics.R
\name{eem_fluorescence_index}
\alias{eem_fluorescence_index}
\title{Calculate the fluorescence index (FI)}
\usage{
eem_fluorescence_index(eem)
}
\arguments{
\item{eem}{An object of class \code{eem}}
}
\value{
A vector containing fluorescence index
}
\description{
Calculate the fluorescence index (FI)
}
\examples{
file <- system.file("extdata/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)

eem_fluorescence_index(eem)
}
\references{
\url{http://doi.wiley.com/10.4319/lo.2001.46.1.0038}
}

