% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_inner_filter_effect}
\alias{eem_inner_filter_effect}
\title{Inner-filter effect correction}
\usage{
eem_inner_filter_effect(eem, absorbance, pathlength = 1)
}
\arguments{
\item{eem}{Either an object of class \code{eem} or a list of \code{eem}.}

\item{absorbance}{A data frame with:

  \describe{ \item{wavelenght}{A numeric vector containing wavelenghts.}
  \item{...}{One or more numeric vectors containing absorbance spectra.}}}

\item{pathlength}{A numeric value indicating the pathlength (in cm) of the
cuvette used for fluorescence measurement. Default is 1 (1cm).}
}
\value{
Either an object of class \code{eem} or a list of \code{eem}.

An object of class \code{eem} containing: \itemize{ \item sample The
  file name of the eem. \item x A matrix with fluorescence values. \item em
  Emission vector of wavelengths. \item ex Excitation vector of wavelengths.
  }
}
\description{
Inner-filter effect correction
}
\details{
The names of \code{absorbance} variables are expected to match those
  of the eems. If the appropriate absorbance spectrum is not found, an
  uncorrected eem will be returned and a warning message will be printed.
}
\examples{
library(eemR)
data("absorbance")
}
\references{
Parker, C. a., & Barnes, W. J. (1957). Some experiments with
  spectrofluorimeters and filter fluorimeters. The Analyst, 82(978), 606.

  \url{http://doi.org/10.1039/an9578200606}
}

