% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_raman_normalisation}
\alias{eem_raman_normalisation}
\title{Title}
\usage{
eem_raman_normalisation(eem, blank)
}
\arguments{
\item{eem}{An object of class \code{eem}}

\item{blank}{An object of class \code{eem}}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
Title
}
\examples{
fluo <- "/home/persican/Desktop/test/eem/sn21.csv"
b <- "/home/persican/Desktop/test/nano.csv"
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)

eem <- eem_read(fluo, ex, em)
blank <- eem_read(b, ex, em)

eem_raman_normalisation(eem, blank)
}

