% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_read.R
\name{eem_read}
\alias{eem_read}
\title{Read excitation-emission fluorecence matrix (eem)}
\usage{
eem_read(file)
}
\arguments{
\item{file}{File name or folder containing fluorescence file(s).}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
Read excitation-emission fluorecence matrix (eem)
}
\details{
Note that \code{em} and \code{ex} are rounded before returned.
}
\examples{
file <- system.file("extdata/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)
}

