% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_read.R
\name{eem_read}
\alias{eem_read}
\title{Read excitation-emission fluorecence matrix (eem)}
\usage{
eem_read(file, ex, em)
}
\arguments{
\item{file}{File name or folder containing fluorescence file.}

\item{ex}{Excitation vector of wavelengths.}

\item{em}{Emission vector of wavelengths.}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
Read excitation-emission fluorecence matrix (eem)
}
\examples{
fluo <- system.file("extdata/eem", "sample1.csv", package = "eem")
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)
eem <- eem_read(fluo, ex, em)
}

