% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove_blank}
\alias{eem_remove_blank}
\title{Blank correction}
\usage{
eem_remove_blank(eem, blank)
}
\arguments{
\item{eem}{An object of class \code{eem}}

\item{blank}{An object of class \code{eem}}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
Blank correction
}
\examples{
fluo <- system.file("extdata/eem", "sample1.csv", package = "eem")
b <- system.file("extdata", "nano.csv", package = "eem")
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)

eem <- eem_read(fluo, ex, em)
blank <- eem_read(b, ex, em)

eem_remove_blank(eem, blank)
}

