% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove_blank}
\alias{eem_remove_blank}
\title{Blank correction}
\usage{
eem_remove_blank(eem, blank)
}
\arguments{
\item{eem}{Either an object of class \code{eem} or a list of \code{eem}.}

\item{blank}{Either an object of class \code{eem} or a list of \code{eem}.}
}
\value{
Either an object of class \code{eem} or a list of \code{eem}.
}
\description{
Blank correction
}
\examples{
# Open the fluorescence eem
file <- system.file("extdata/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)

plot(eem)

# Open the blank eem
file <- system.file("extdata", "nano.csv", package = "eemR")
blank <- eem_read(file)

plot(blank)

# Remove the blank
eem <- eem_remove_blank(eem, blank)

plot(eem)
}

