% Generated by roxygen2 (5.0.0): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove_scattering}
\alias{eem_remove_scattering}
\title{Remove Raman and Rayleigh scattering}
\usage{
eem_remove_scattering(eem, type, order = 1, width)
}
\arguments{
\item{eem}{Either an object of class \code{eem} or a list of \code{eem}.}

\item{type}{A string, either "raman" or "rayleigh"}

\item{order}{A integer number, either 1 (first order) or 2 (second order)}

\item{width}{Slit width in nm for the cut}
}
\value{
Either an object of class \code{eem} or a list of \code{eem}.
}
\description{
Remove Raman and Rayleigh scattering
}
\examples{
# Open the fluorescence eem
file <- system.file("extdata/eem", "sample1.csv", package = "eemR")
eem <- eem_read(file)

plot(eem)

# Remove the scattering
eem <- eem_remove_scattering(eem = eem, type = "raman", order = 1, width = 10)

plot(eem)
}

