% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{eem_remove_scattering}
\alias{eem_remove_scattering}
\title{Remove Raman and Rayleigh scattering}
\usage{
eem_remove_scattering(eem, type, order = 1, width)
}
\arguments{
\item{eem}{eem An object of class \code{eem}}

\item{type}{A string, either "raman" or "rayleigh"}

\item{order}{A integer number, either 1 (first order) or 2 (second order)}

\item{width}{Slit width in nm for the cut}
}
\value{
An object of class \code{eem} containing:
\itemize{
 \item sample The file name of the eem.
 \item x A matrix with fluorescence values.
 \item em Emission vector of wavelengths.
 \item ex Excitation vector of wavelengths.
}
}
\description{
Remove Raman and Rayleigh scattering
}
\examples{
fluo <- "/home/persican/Desktop/test/eem/sn21.csv"
b <- "/home/persican/Desktop/test/nano.csv"
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)

eem <- eem_read(fluo, ex, em)

eem_remove_scattering(eem = eem, type = "raman", order = 1, width = 10)
}

