% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eemR-package.R
\docType{package}
\name{eemR-package}
\alias{eemR}
\alias{eemR-package}
\title{eemR: Tools for Pre-Processing Emission-Excitation-Matrix (EEM) Fluorescence
    Data}
\description{
Provides various tools for preprocessing Emission-Excitation-Matrix (EEM) for Parallel Factor Analysis (PARAFAC). Different
  methods are also provided to calculate common metrics such as humification index and fluorescence index.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PMassicotte/eemR}
  \item Report bugs at \url{https://github.com/PMassicotte/eemR/issues}
}

}
\author{
\strong{Maintainer}: Philippe Massicotte \email{pmassicotte@hotmail.com}

}
\keyword{internal}
