% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{plot.eem}
\alias{plot.eem}
\title{Surface plot of eem}
\usage{
\method{plot}{eem}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eemlist}.}

\item{...}{Extra arguments for \code{image.plot}.}
}
\description{
Surface plot of eem
}
\examples{
file <- system.file("extdata/cary/eem/", "sample1.csv", package = "eemR")
eem <- eem_read(file)

plot(eem)
}

