% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{plot.eem}
\alias{plot.eem}
\title{Surface plot of eem}
\usage{
\method{plot}{eem}(x, ...)
}
\arguments{
\item{x}{An object of class \code{eem}}

\item{...}{Extra arguments for \code{image.plot}}
}
\description{
Surface plot of eem
}
\examples{
fluo <- "/home/persican/Desktop/test/eem/sn21.csv"
ex <- seq(220, 450, by = 5)
em <- seq(230, 600, by = 2)
eem <- eem_read(fluo, ex, em)
plot(eem)
}

