% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eem_utils.R
\name{summary.eem}
\alias{summary.eem}
\title{Display summary of an eem object}
\usage{
\method{summary}{eem}(object, ...)
}
\arguments{
\item{object}{An object of class \code{eem}}

\item{...}{Extra arguments}
}
\description{
Display summary of an eem object
}
\examples{
fluo <- system.file("extdata/eem", "sample1.csv", package = "eemR")
eem <- eem_read(fluo)
summary(eem)
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/0304420395000623}
}

